feat(nav): add nav bar with site header and logout button
